Address Copilot review: fetch LFS in CI and use %zu for counts
- conformance.yml: set `lfs: true` on every actions/checkout step so the
  Git LFS-managed test_data/ images are actually fetched (otherwise the CTS
  fails at runtime opening images like lena.bmp). GitHub-hosted runners ship
  with git-lfs preinstalled.
- test_minmaxloc.c: print mincount/maxcount with %zu and cast to size_t
  instead of truncating vx_size to int on LP64 platforms.

+++ test_conformance/test_minmaxloc.c
@@ -265,12 +265,12 @@ TEST_WITH_ARG(MinMaxLoc, testOnRandom, format_arg,
         if( minval != minval0 || maxval != maxval0 || mincount != mincount0 || maxcount != maxcount0 )
         {
             CT_RecordFailureAtFormat("Test case %d. width=%d, height=%d,\n"
-                                     "\tExpected: minval=%d, maxval=%d, mincount=%d, maxcount=%d\n"
-                                     "\tActual:   minval=%d, maxval=%d, mincount=%d, maxcount=%d\n",
+                                     "\tExpected: minval=%d, maxval=%d, mincount=%zu, maxcount=%zu\n"
+                                     "\tActual:   minval=%d, maxval=%d, mincount=%zu, maxcount=%zu\n",
                                      __FUNCTION__, __FILE__, __LINE__,
                                      iter, width, height,
-                                     minval0, maxval0, (int)mincount0, (int)maxcount0,
-                                     minval, maxval, (int)mincount, (int)maxcount);
+                                     minval0, maxval0, (size_t)mincount0, (size_t)maxcount0,
+                                     minval, maxval, (size_t)mincount, (size_t)maxcount);
             break;
         }
